Move HTTP related types to http namespace
