+++ sdk/appconfiguration/Azure.ApplicationModel.Configuration/src/ConfigurationClientModelFactory.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using Azure.Core.Http;
 
 namespace Azure.ApplicationModel.Configuration
 {
