+++ sdk/appconfiguration/Azure.ApplicationModel.Configuration/src/ConfigurationClient_private.cs
@@ -11,6 +11,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure.Core;
+using Azure.Core.Http;
 using Azure.Core.Pipeline;
 
 namespace Azure.ApplicationModel.Configuration
