+++ sdk/appconfiguration/Azure.ApplicationModel.Configuration/src/ConfigurationSetting.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using Azure.Core;
+using Azure.Core.Http;
 
 namespace Azure.ApplicationModel.Configuration
 {
