+++ sdk/appconfiguration/Azure.ApplicationModel.Configuration/src/ConfigurationSettingParser.cs
@@ -9,6 +9,7 @@ using System.IO;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
+using Azure.Core.Http;
 
 namespace Azure.ApplicationModel.Configuration
 {
