+++ sdk/appconfiguration/Azure.ApplicationModel.Configuration/tests/ConfigurationMockTests.cs
@@ -10,6 +10,7 @@ using System.Text;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
+using Azure.Core.Http;
 using NUnit.Framework;
 
 namespace Azure.ApplicationModel.Configuration.Tests
