+++ sdk/appconfiguration/Azure.ApplicationModel.Configuration/tests/ConfigurationSettingTests.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using Azure.Core;
+using Azure.Core.Http;
 
 namespace Azure.ApplicationModel.Configuration.Tests
 {
