+++ sdk/core/Azure.Core/src/Http/HttpClientTransport.cs
