+++ sdk/core/Azure.Core/src/Http/HttpHeader.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace Azure.Core.Pipeline
+namespace Azure.Core.Http
 {
     public readonly struct HttpHeader : IEquatable<HttpHeader>
     {
