+++ sdk/core/Azure.Core/src/Http/RequestHeaders.cs
