+++ sdk/core/Azure.Core/src/Pipeline/Policies/LoggingPolicy.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure.Core.Diagnostics;
+using Azure.Core.Http;
 
 namespace Azure.Core.Pipeline.Policies
 {
