+++ sdk/core/Azure.Core/tests/ETagTests.cs
@@ -1,10 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-using System.Collections.Generic;
-using System.Text;
-using NUnit;
+using Azure.Core.Http;
 using NUnit.Framework;
 
 namespace Azure.Core.Tests
