+++ sdk/core/Azure.Core/tests/HeadersTests.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Text;
 using System.Text.RegularExpressions;
+using Azure.Core.Http;
 using Azure.Core.Pipeline;
 using Azure.Core.Testing;
 using NUnit.Framework;
