+++ sdk/storage/Azure.Storage.Blobs/src/Models/ReleasedObjectInfo.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.ComponentModel;
+using Azure.Core.Http;
 
 namespace Azure.Storage.Blobs.Models
 {
