+++ sdk/storage/Azure.Storage.Common/swagger/Generator/src/types.ts
@@ -22,7 +22,7 @@ export function getName(model: IModelType, readonly?: boolean, isParam?: boolean
             case 'binary':
             case 'file': return 'System.IO.Stream';
             case 'url': return 'System.Uri';
-            case 'etag': return 'Azure.ETag';
+            case 'etag': return 'Azure.Core.Http.ETag';
             case 'date': return 'System.DateTime';
             case 'date-time':
             case 'date-time-8601':
@@ -170,7 +170,7 @@ export function convertFromString(expr: string, model: IModelType, service: ISer
         case 'date-time-rfc1123':
             return `System.DateTimeOffset.Parse(${expr}, System.Globalization.CultureInfo.InvariantCulture)`;
         case 'etag':
-            return `new Azure.ETag(${expr})`;
+            return `new Azure.Core.Http.ETag(${expr})`;
         case 'url':
             return `new System.Uri(${expr})`;
         case 'byte':
